backend/end-points/requirements.txt
# FastAPI and server requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Google ADK and AI requirements
google-adk>=1.14.0
google-generativeai>=0.8.5
google-cloud-aiplatform

# Utility requirements
python-dotenv>=1.0.0
pathlib

# Audio processing (if needed)
numpy